backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
websockets==13.1
aiosqlite==0.20.0
